add bisection search for a target spectral compression.
